Move base rest handler class
