+++ x-pack/plugin/ent-search/src/main/java/org/elasticsearch/xpack/application/search/action/RestDeleteSearchApplicationAction.java
@@ -14,13 +14,14 @@ import org.elasticsearch.rest.Scope;
 import org.elasticsearch.rest.ServerlessScope;
 import org.elasticsearch.rest.action.RestToXContentListener;
 import org.elasticsearch.xpack.application.EnterpriseSearch;
+import org.elasticsearch.xpack.application.EnterpriseSearchBaseRestHandler;
 
 import java.util.List;
 
 import static org.elasticsearch.rest.RestRequest.Method.DELETE;
 
 @ServerlessScope(Scope.PUBLIC)
-public class RestDeleteSearchApplicationAction extends SearchApplicationRestHandler {
+public class RestDeleteSearchApplicationAction extends EnterpriseSearchBaseRestHandler {
     public RestDeleteSearchApplicationAction(XPackLicenseState licenseState) {
         super(licenseState);
     }
