+++ x-pack/plugin/ent-search/src/main/java/org/elasticsearch/xpack/application/search/action/RestGetSearchApplicationAction.java
@@ -14,13 +14,14 @@ import org.elasticsearch.rest.Scope;
 import org.elasticsearch.rest.ServerlessScope;
 import org.elasticsearch.rest.action.RestToXContentListener;
 import org.elasticsearch.xpack.application.EnterpriseSearch;
+import org.elasticsearch.xpack.application.EnterpriseSearchBaseRestHandler;
 
 import java.util.List;
 
 import static org.elasticsearch.rest.RestRequest.Method.GET;
 
 @ServerlessScope(Scope.PUBLIC)
-public class RestGetSearchApplicationAction extends SearchApplicationRestHandler {
+public class RestGetSearchApplicationAction extends EnterpriseSearchBaseRestHandler {
     public RestGetSearchApplicationAction(XPackLicenseState licenseState) {
         super(licenseState);
     }
